@stephenh 2022-11-26
Hey @gwer , this is great!

I must admit I went too long without paying attention to performance, and then recently got lucky that just switching from prettier to dprint gave a huge performance gain, and then I had stopped there.

I'll release a new version of this and then we can bump ts-poet in ts-proto and see how it works over there (i.e. the ts-proto integration tests can serve as another regression test of this change). But so far everything looks good to me!