@stephenh 2022-11-26
I manually published ts-poet 6.2.0 with this change. Thanks again!